+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add `loop` action to multiply several actions. ([@palkan][])
+
 ## 1.0.2 (2023-01-12)
 
 - Fix adding transpiled files to releases. ([@palkan][])